python/sedonadb/pyproject.toml
[build-system]
requires = ["maturin>=1.8,<2.0"]
build-backend = "maturin"

[project]
name = "sedonadb"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Rust",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dynamic = ["version"]

[project.optional-dependencies]
test = ["pytest", "adbc-driver-manager[dbapi]", "pandas", "geopandas", "geoarrow-pyarrow"]

[tool.maturin]
features = ["pyo3/extension-module"]
module-name = "sedonadb._lib"
python-source = "python"
